requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.20
langchain-google-genai==1.0.1
langchain-community==0.0.38
firebase-admin==6.4.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6